requirements.txt
django==1.11
djangorestframework
model_mommy=1.5.0